+++ src/Helper.js
@@ -38,7 +38,7 @@ Helper.prototype.getNoteFrequencies = function() {
  * @param  {Float}       velocity Like Midi velocity but high res float of range [0, 1]
  * @return {Object}               A note event
  */
-Helper.prototype.createNoteEvent = function(uid, tone, velocity) {
+Helper.prototype.createNoteEvent = function(uid, tone, velocity, duration) {
   var noteNum, evt;
 
   if (typeof tone === 'number') {
@@ -52,7 +52,8 @@ Helper.prototype.createNoteEvent = function(uid, tone, velocity) {
     msg: {
       type: 'note',
       value: noteNum,
-      velocity: velocity
+      velocity: velocity,
+      duration: duration
     }
   };
 
@@ -62,6 +63,9 @@ Helper.prototype.createNoteEvent = function(uid, tone, velocity) {
     }
     if (velocity < 0 || velocity > 1) {
       throw new Error('Velocity out of bounds: ' + velocity);
+    }
+    if (typeof duration !== 'number' && typeof duration !== 'undefined') {
+      throw new TypeError('Duration if defined must be of type number.');
     }
     return evt;
   } else {